[feat] helmrepo chart 정렬 위해서 repository get 부분 수정 및 정렬 추가
